% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grid_lasso.R
\name{cv_grid_lasso}
\alias{cv_grid_lasso}
\title{Computes cross-validated solutions for continuous response model}
\usage{
cv_grid_lasso(
  x,
  y,
  K = 5,
  var_order = NULL,
  lambda = NULL,
  nlambda = 100L,
  lambda.min.ratio = ifelse(n < p, 0.01, 1e-04),
  grid.size = p,
  thresh = 1e-10,
  maxit = 1e+05,
  sparse = TRUE,
  mc.cores = 1,
  return.full.beta = FALSE,
  silent = TRUE,
  early.stopping = TRUE,
  early.stopping.factor = 0.5,
  fold_assign = NULL,
  missing.data = F,
  psd.method = "enet"
)
}
\arguments{
\item{x}{Design matrix, n x p.}

\item{y}{Vector of responses, length n.}

\item{K}{Number of folds for cross-validation. Must be at least 2}

\item{var_order}{For user-specified ordering of variables. Indices start at 0, start with least important variable and end with most. By default order will be induced from scaling of columns in design matrix}

\item{lambda}{For user-specified sequence of tuning parameter lambda}

\item{nlambda}{Length of automatically generated sequence of tuning parameters lambda}

\item{lambda.min.ratio}{Ratio of max/min lambda for automatically generated sequence of tuning parameters lambda}

\item{grid.size}{Number of subsets of variables for which a solution path will be computed for}

\item{thresh}{Convergence threshold for coordinate descent for difference in objective values between successive iterations}

\item{maxit}{Maximum number of iterations for coordinate descent routine}

\item{sparse}{Whether to use sparse matrices in computation (setting FALSE recommended for advanced users only)}

\item{mc.cores}{Number of cores to be made available for computing the cross-validation estimates in parallel}

\item{return.full.beta}{Return the entire solution path for the chosen variable subset, as opposed to only the estimate for estimated optimal lambda}

\item{silent}{Suppress some text to console}

\item{early.stopping}{Whether square-root lasso condition for early stopping along lambda path should be used}

\item{early.stopping.factor}{Factor of correction in square-root lasso early stopping criterion}

\item{fold_assign}{For user-specified vector of assignment of folds for cross-validation. Must be of the form of integer vector with entries in 1 , ... , K.}

\item{missing.data}{If TRUE then will use (slower) procedure that corrects for missing data}

\item{psd.method}{The way that the gram matrix is made positive semidefinite. By default an elastic net term, alternatives are "coco" for CoCoLasso}
}
\value{
A list of objects: 
\itemize{
\item mu -- estimated intercept
\item beta -- coefficient estimate
\item beta.full -- full solution path (only returned if return.full.beta = TRUE)
\item lambda -- Vector of values for lambda used
\item cv -- matrix of cross-validation error for each value of lambda and gridpoint
\item col.means -- vector of column means in unstandardized design matrix; important for making predictions on new data.
}
}
\description{
Cross-validation wrapper for grid_lasso that computes solutions, selects and fits the optimal model.
}
\examples{
set.seed(1)
X = matrix(0, 50, 500)
Z = matrix(0, 10, 500)
betavec = c(rep(1,5),rep(0,495))
X[ , 1:5 ] = matrix(rnorm(250), 50, 5)
Z[ , 1:5 ] = matrix(rnorm(50), 10, 5)
Y = X \%*\% betavec
Y = Y + rnorm(50)
X = X + matrix(rnorm(50*500), 50, 500)
mod1 = cv_grid_lasso(X, Y, grid.size = 50)
predict(mod1, Z)

}
