% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grid_lasso_logistic.R
\name{cv_grid_lasso_logistic}
\alias{cv_grid_lasso_logistic}
\title{Computes cross-validated solutions for binary response model}
\usage{
cv_grid_lasso_logistic(
  x = NULL,
  y,
  K = 5,
  var_order = NULL,
  lambda = NULL,
  nlambda = 100L,
  grid.size = p,
  lambda.min.ratio = ifelse(n < p, 0.01, 1e-04),
  thresh = 1e-10,
  maxit = 1e+05,
  mc.cores = 1,
  return.full.beta = FALSE,
  silent = TRUE,
  fold_assign = NULL
)
}
\arguments{
\item{x}{Design matrix, n x p.}

\item{y}{Vector of responses, length n.}

\item{K}{Number of folds for cross-validation. Must be at least 2}

\item{var_order}{For user-specified ordering of variables. Indices start at 0, start with least important variable and end with most. By default order will be induced from scaling of columns in design matrix}

\item{lambda}{For user-specified sequence of tuning parameter lambda}

\item{nlambda}{Length of automatically generated sequence of tuning parameters lambda}

\item{grid.size}{Number of subsets of variables for which a solution path will be computed for}

\item{lambda.min.ratio}{Ratio of max/min lambda for automatically generated sequence of tuning parameters lambda}

\item{thresh}{Convergence threshold for coordinate descent for difference in objective values between successive iterations}

\item{maxit}{Maximum number of iterations for coordinate descent routine}

\item{mc.cores}{Number of cores to be made available for computing the cross-validation estimates in parallel}

\item{return.full.beta}{Return the entire solution path for the chosen variable subset, as opposed to only the estimate for estimated optimal lambda}

\item{silent}{Suppress some text to console}

\item{fold_assign}{For user-specified vector of assignment of folds for cross-validation. Must be of the form of integer vector with entries in 1 , ... , K.}
}
\value{
A glmnet model object, with some additional attributes:
\itemize{
\item best -- an index denoting which point on the path of lambda values is estimated optimal
\item cv -- matrix of cross-validation error for each value of lambda and gridpoint
}
}
\description{
Cross-validation wrapper for grid_lasso_logistic that computes solutions, selects and fits the optimal model.
}
\examples{
set.seed(1)
X = matrix(0, 50, 500)
betavec = c(rep(1,5),rep(0,495))
X[ , 1:5 ] = matrix(rnorm(250), 50, 5)
Y = as.vector(X \%*\% betavec)
Y = Y + rnorm(50)
Y = as.numeric(Y >= mean(Y))
X = X + matrix(rnorm(50*500), 50, 500)
mod1 = cv_grid_lasso_logistic(X, Y, grid.size = 25)

}
