% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grid_ridge.R
\name{cv_grid_ridge}
\alias{cv_grid_ridge}
\title{Computes cross-validated solution for ridge regression}
\usage{
cv_grid_ridge(
  x,
  y,
  K = 5,
  var_order = NULL,
  lambda = NULL,
  nlambda = 100L,
  lambda.min.ratio = ifelse(n < p, 0.01, 1e-04),
  grid.size = p,
  lambda.mult = 1e+05,
  fold_assign = NULL
)
}
\arguments{
\item{x}{Design matrix, n x p.}

\item{y}{Vector of responses, length n.}

\item{K}{Number of folds for cross-validation. Must be at least 2}

\item{var_order}{For user-specified ordering of variables. Indices start at 0, start with least important variable and end with most. By default order will be induced from scaling of columns in design matrix}

\item{lambda}{For user-specified sequence of tuning parameter lambda}

\item{nlambda}{Length of automatically generated sequence of tuning parameters lambda}

\item{lambda.min.ratio}{Ratio of max/min lambda for automatically generated sequence of tuning parameters lambda}

\item{grid.size}{Number of subsets of variables for which a solution path will be computed for}

\item{lambda.mult}{Scales the sequence of lambda by a constant}

\item{fold_assign}{For user-specified vector of assignment of folds for cross-validation. Must be of the form of integer vector with entries in 1 , ... , K.}
}
\value{
A list of objects:
\itemize{
\item mu -- estimated intercept
\item beta -- coefficient estimate
\item cv -- a matrix of errors for the models (grid.size times nlambda)
\item lambda -- sequence of lambda values used
}
}
\description{
Cross-validation wrapper for grid_ridge that computes solutions, selects and fits the optimal model.
}
\examples{
set.seed(1)
X = matrix(0, 50, 100)
betavec = c(rep(1,5),rep(0,95))
X[ , 1:5 ] = matrix(rnorm(250), 50, 5)
Y = as.vector(X \%*\% betavec)
Y = Y + rnorm(50)
X = X + matrix(rnorm(50*100), 50, 100)
mod1 = cv_grid_ridge(X, Y, grid.size = 50)

}
