% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_ridge.R
\name{grid_ridge}
\alias{grid_ridge}
\title{Computes grid of predictions for ridge regression}
\usage{
grid_ridge(
  x,
  z,
  y,
  yz,
  var_order = NULL,
  lambda = NULL,
  nlambda = 100,
  lambda.min.ratio = 1e-05,
  grid.size = p,
  lambda.mult = 1e+05,
  errors_mean = TRUE
)
}
\arguments{
\item{x}{design matrix for training}

\item{z}{design matrix for testing}

\item{y}{response for training}

\item{yz}{response for testing}

\item{var_order}{For user-specified ordering of variables. Indices start at 0, start with least important variable and end with most. By default order will be induced from scaling of columns in design matrix}

\item{lambda}{For user-specified sequence of tuning parameter lambda}

\item{nlambda}{Length of automatically generated sequence of tuning parameters lambda}

\item{lambda.min.ratio}{Ratio of max/min lambda for automatically generated sequence of tuning parameters lambda}

\item{grid.size}{Number of subsets of variables for which a solution path will be computed for}

\item{lambda.mult}{Scales the sequence of lambda by a constant}

\item{errors_mean}{Controls whether MSPE or SPE (without dividing by sample size) is returned. Used only for within cv_grid_ridge}
}
\value{
A list of objects:
\itemize{
\item errors -- a matrix of errors for the models (grid.size times nlambda)
\item lambda -- sequence of lambda values used
}
}
\description{
Computes prediction errors for a grid (over lambda and variable subsets) of ridge regression models
}
\examples{
set.seed(1)
X = matrix(0, 50, 500)
Z = matrix(0, 10, 500)
betavec = c(rep(1,5),rep(0,495))
X[ , 1:5 ] = matrix(rnorm(250), 50, 5)
Z[ , 1:5 ] = matrix(rnorm(50), 10, 5)
YZ = as.vector(Z \%*\% betavec)
Y = as.vector(X \%*\% betavec)
Y = Y + rnorm(50)
X = X + matrix(rnorm(50*500), 50, 500)
mod1 = grid_ridge(X, Z, Y, YZ, grid.size = 50)

}
