% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_ridge_lambda.R
\name{grid_ridge_lambda}
\alias{grid_ridge_lambda}
\title{Computes grid of predictions for ridge regression for a specific value of lambda}
\usage{
grid_ridge_lambda(
  x,
  z,
  y,
  yz,
  var_order = NULL,
  lambda,
  x1x1inv = NULL,
  grid.size = p,
  errors_mean = TRUE
)
}
\arguments{
\item{x}{design matrix for training}

\item{z}{design matrix for testing}

\item{y}{response for training}

\item{yz}{response for testing}

\item{var_order}{For user-specified ordering of variables. Indices start at 0, start with least important variable and end with most. By default order will be induced from scaling of columns in design matrix}

\item{lambda}{Value of lambda to be used for the calculation}

\item{x1x1inv}{If this is already computed then it can be passed here. Only for use within grid_ridge}

\item{grid.size}{Number of subsets of variables for which a solution path will be computed for}

\item{errors_mean}{Controls whether MSPE or SPE (without dividing by sample size) is returned. Used only for within cv_grid_ridge}
}
\value{
A vector of errors of length grid.size
}
\description{
Computes prediction errors for a grid of ridge regression models for a specified value of lambda
}
