% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv_grid_lasso.R
\name{predict.cv_grid_lasso}
\alias{predict.cv_grid_lasso}
\title{Predictions for continuous response model}
\usage{
\method{predict}{cv_grid_lasso}(object, newx, l = NULL, missing.data = FALSE, psd.method = "enet", ...)
}
\arguments{
\item{object}{Object as returned by grid_lasso or cv_grid_lasso, provided return.list == T}

\item{newx}{Design matrix for new observations}

\item{l}{Specify point on solution path (optional)}

\item{missing.data}{If TRUE then will use (slower) procedure that corrects for missing data, and will only return a (constant shifted) mean square error rather than actual predictions}

\item{psd.method}{The way that the gram matrix is made positive semidefinite. By default an elastic net term, alternatives are "coco" for CoCoLasso}

\item{...}{Additional arguments to pass to other \code{predict} methods}
}
\value{
Vector of predictions (or constant-shifted error if missing.data = TRUE)
}
\description{
Returns predicted response from grid lasso model with new data
}
